backend/rag_core/requirements.txt
# Core ML/DL
torch
transformers
bitsandbytes
accelerate
scikit-learn
sentence-transformers

# LangChain & Google API
langchain
langchain-core
langchain-community
langchain-google-genai
google-generativeai
python-dotenv

# Data & DB
chromadb
PyYAML
tqdm
pandas
